feat: Add distance feature
Player can see how far they've run. Points are allocated as you run.

To accomplish this, we add fractional points, and improve score component to only report
full integer version of the score. The score component is also modified so it can be used
by the distance tracker as well.