feat: Enable player variable bouncing heights
If the player bounces off an enemy, he will do a short bounce and that's it. But if he is
holding down the jump button, he will do a high bounce, and will have the ability to
control the height of his bounce by whether he continues to hold down the jump button,
because it's actually a continuation of a variable jump.